+++ pyincore/analyses/equitymetric/equitymetricutil.py
@@ -9,41 +9,41 @@ import pandas as pd
 
 class EquityMetricUtil:
     @staticmethod
-    def prepare_svi_as_division_decision(merged_gdf):
+    def prepare_svi_as_division_decision(hua_df):
         """
         socially vulnerability as division decision variable which is a binary variable associated with each household
         used to group it into two groups
         Args:
-            merged_gdf:
+            hua_df:
 
         Returns:
 
         """
         # Add variable to indicate if high socially vulnerability for metric's computation
-        median_income = merged_gdf["randincome"].median()
+        median_income = hua_df["randincome"].median()
 
-        condition1 = merged_gdf["randincome"] <= median_income
-        condition2 = merged_gdf["ownershp"] == 2
-        condition3 = merged_gdf["race"] != 1
-        condition4 = merged_gdf["hispan"] != 0
+        condition1 = hua_df["randincome"] <= median_income
+        condition2 = hua_df["ownershp"] == 2
+        condition3 = hua_df["race"] != 1
+        condition4 = hua_df["hispan"] != 0
 
-        merged_gdf["SVI"] = condition1 & condition2 & condition3 & condition4
-        merged_gdf["SVI"] = (merged_gdf["SVI"]).astype(int)
+        hua_df["SVI"] = condition1 & condition2 & condition3 & condition4
+        hua_df["SVI"] = (hua_df["SVI"]).astype(int)
 
-        return merged_gdf
+        return hua_df
 
     @staticmethod
-    def prepare_return_time_as_scarce_resrouce(return_df):
+    def prepare_return_time_as_scarce_resource(return_df):
         return_sequence = return_df.iloc[:, 4:94]
         # add return time to the scarce resource dataset
-        time_to_return = EquityMetricUtil.time_to_return(return_sequence)
+        time_to_return = EquityMetricUtil._time_to_return(return_sequence)
         return_df["Return Time"] = pd.to_numeric(time_to_return)
         return_df["scarce_resource"] = 91 - return_df["Return Time"]
 
         return return_df
 
     @staticmethod
-    def time_to_return(return_sequence):
+    def _time_to_return(return_sequence):
         # now create a for loop to determine the time for each row
         time_to_return = []
         for i in range(0, return_sequence.shape[0]):